+++ internal/app/app_impl.go
@@ -53,10 +53,12 @@ type _AppImpl struct {
 	inprocFlag *InternalInprocFlag
 
 	injector            di.PlainInjector
+	moduleInjector      di.PlainInjector
 	componentManagers   []ComponentManager
 	components          []Component
 	componentLifecycles []ComponentLifecycle
 	modules             []Module
+	hooks               HookAdder
 
 	httpServer *http.Server
 	httpWG     sync.WaitGroup
@@ -162,10 +164,12 @@ func (a *_AppImpl) Start() {
 	vpre.Check(a.lifecycleState == appLifecycleStateNew, "application already stopped")
 	a.lifecycleState = appLifecycleStateStarted
 
+	a.spec.InitHooks(&a.hooks)
 	a.initLinking()
 	a.initInjector()
 	a.initComponents()
 	a.initModules()
+	a.prepareHooks()
 	a.initServers()
 
 	err := a.beforeAppStart()
@@ -258,6 +262,9 @@ func (a *_AppImpl) shouldEnableConsole() bool {
 }
 
 func (a *_AppImpl) beforeAppStart() error {
+	if err := invokeHooks(a.hooks.beforeStart, false); err != nil {
+		return err
+	}
 	for _, componentHook := range a.componentLifecycles {
 		if err := componentHook.BeforeAppStart(); err != nil {
 			return err
@@ -278,9 +285,11 @@ func (a *_AppImpl) afterAppStart() {
 	for _, module := range a.modules {
 		module.AfterAppStart()
 	}
+	invokeHooks(a.hooks.afterStart, false)
 }
 
 func (a *_AppImpl) beforeAppStop() {
+	invokeHooks(a.hooks.beforeStop, true)
 	for i := range a.modules {
 		a.modules[len(a.modules)-1-i].BeforeAppStop()
 	}
@@ -296,6 +305,7 @@ func (a *_AppImpl) afterAppStop() {
 	for i := range a.componentLifecycles {
 		a.componentLifecycles[len(a.componentLifecycles)-1-i].AfterAppStop()
 	}
+	invokeHooks(a.hooks.afterStop, true)
 }
 
 func (a *_AppImpl) registerApp() {